+++ zhuanlan.py
@@ -7,6 +7,7 @@ import json
 from mako.template import Template
 from session import session
 from article import Article
+import os.path
 
 
 class ZhuanLan:
@@ -26,6 +27,9 @@ class ZhuanLan:
         """ 下载 url 指定的图片并保存到本地，并返回本地路径"""
         # 提取 url 中图片的名称
         name = url[url.rfind('/')+1:]
+        # 如果本地已经有该图片就不再下载，节省时间和流量
+        if os.path.isfile('./img/' + name):
+            return './img/' + name
         binary_content = self.s.get(url=url).content
         path = self.img_path + '/' + name
         with open(file=path, mode='wb') as f:
